Added the HeatMap class as well as a testing scene (FaceHeatMapTest).
The way it works: Feeding the HeatMap a Texture2D and a FaceAttribute (currently implemented Gender and Age) will create a new
Texture2D with colored squares placed on top of the faces. So, for example, if you select Gender, then male/female/genderless people
will get different color boxes. You can also de-saturate the image for a clearer view.

Going forward, this needs to be replaced by shaders that do the same work and are portable to different divices.